changed from int commands to floats in ch01 exercise

+++ ch01/exercises/simple.py
@@ -10,19 +10,19 @@ print(0 % 10)
 print(10 / 15) #the answer is neverending, but in the program it eventually ends
 
 
-rate = int(input("Input ER from Euro to USD: "))
+rate = float(input("Input ER from Euro to USD: "))
 #rate variable that asks for an exchange rate
 print("The Exchange Rate is ",[rate])
 #prints rate variable to user
-amount = int(input("How much currency to exchange: "))
+amount = float(input("How much currency to exchange: "))
 #input an amount to convert from user
 print("The input is $",[amount])
 #print the amount variable for user
-total = int(rate * amount)
+total = float(rate * amount)
 #creates total variable which is the converstion of the amount inputted and the rate
 print("USD Received before fees: ",[total])
 #prints the total variable for user
-result = int(total - 3)
+result = float(total - 3)
 #creates result variable which subtracts the fee from total variable
 print("USD Received after $3 fee: ",[result])
 #prints the result variable for the user as a final result of the code
